SQL join
Q 1. What is the total quantity of products sold in the month of May 2022?

Q 2. What is the total revenue generated by the category "Electronics" in the month of December 2022?

Q 3. Write a query to return the name and email of all customers who have placed an order in the month 
    of May and have ordered a product with a price greater than $100. Which of the following options returns the correct result?

Q 4. Write a query to return the name and address of all customers who have ordered a product from the
    category 'Electronics' and whose order status is 'Delivered'. Which of the following options returns the correct result?

Q 5.Write a query to return the name and email of all customers who have placed an order, and also include customers
    who haven't placed any orders yet. Which of the following options returns the correct result?

Q 6. Write a query to return the product name and price of all products, including products that have 
    no orders placed yet. Which of the following options returns the correct result?

Q 7. Write a query to return the name and email of all customers who have placed an order for a product with a
    price greater than $100. Include customers who have not placed any orders. How many entries of such details are there?

Q 8. Write a query to return the name and address of all customers who have placed an order for a product in the 
    category 'Electronics', including customers who have not placed any orders. How many such rows are there?

Q 9.Write a query to return all department names along with the number of employees in each department, 
    including departments with no employees. Which of the following options correctly identifies the departments with no employees?

Q 10. Write a query to return the names of all customers who have not placed an order?

Q 11. Write a query to return the customer name and order date for all orders, including 
    those with no corresponding customer or product data. Which of the following options returns the correct result?

Q 12. Write a query to cross join the "customers" and "products" tables. If the "customers" table contains 50 records 
    and the "products" table contains 100 records, how many records will be returned in the result set?

Q 13.Write a query to cross join the "orders" and "order_details" tables. If the "orders" table contains 500 records 
    and the "order_details" table contains 20 records, how many records will be returned in the result set?

Q 14. Write a query to perform a cross join between the "customers" and "orders" tables. 
    How many records will be returned by this query where the "order_status" is "Delivered"?

Q 15. Write a query to perform a cross join between the "products" and "categories" tables.
    How many records will be returned by this query where the "category_name" is "Electronics"?